Fix CodeQL alerts (#278)
* Fix CodeQL alerts

Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/33
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/32
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/31
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/30
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/29
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/8
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/7
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/38
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/37
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/36
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/35
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/34
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/28
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/27
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/26
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/25
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/24
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/23
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/22
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/21
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/20
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/19
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/18
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/17
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/16
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/15
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/14
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/13
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/12
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/11
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/10
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/9
Fixes https://github.com/Open-KO/KnightOnline/security/code-scanning/6

* Correct CodeQL init language identifier

+++ deps/ZipArchive/ZipArchive.cpp
@@ -854,7 +854,7 @@ void CZipArchive::DeleteFiles(CStringArray &aNames, bool bCaseSensitive)
 {
 	CWordArray indexes;
 	
-	for (WORD i = 0; i < GetNoEntries(); i++)
+	for (WORD i = 0; i < static_cast<WORD>(GetNoEntries()); i++)
 	{
 		CZipFileHeader fh;
 		GetFileInfo(fh, i);
